EventListener: Use [JsonSourceGenerationOptions] & cleanup
