+++ src/MudBlazor/Services/EventManager/EventListener.cs
@@ -2,74 +2,19 @@
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using System.Linq;
 using System.Text.Json;
-using System.Threading.Tasks;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.JSInterop;
 
 namespace MudBlazor
 {
-    public interface IEventListenerFactory
-    {
-        IEventListener Create();
-    }
-
-    public class EventListenerFactory : IEventListenerFactory
-    {
-        private readonly IServiceProvider _provider;
-
-        public EventListenerFactory(IServiceProvider provider)
-        {
-            _provider = provider;
-        }
-
-        public IEventListener Create() =>
-            new EventListener(_provider.GetRequiredService<IJSRuntime>());
-    }
-
-    public interface IEventListener : IAsyncDisposable
-    {
-        /// <summary>
-        /// Listing to a javascript event
-        /// </summary>
-        /// <typeparam name="T">The type of the event args for instance MouseEventArgs for mousemove</typeparam>
-        /// <param name="eventName">Name of the DOM event without "on"</param>
-        /// <param name="elementId">The value of the id field of the DOM element</param>
-        /// <param name="projectionName">The name of a JS function (relative to window) that used to project the event before it is send back to .NET. Can be null, if no projection is needed </param>
-        /// <param name="throotleInterval">The delay between the last time the event occurred and the callback is fired. Set to zero, if no delay is requested</param>
-        /// <param name="callback">The method that is invoked, if the DOM element is fired. Object will be of type T</param>
-        /// <returns>A unique identifier for the event subscription. Should be used to cancel the subscription</returns>
-        Task<Guid> Subscribe<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicProperties)] T>(string eventName, string elementId, string projectionName, int throotleInterval, Func<object, Task> callback);
-
-        /// <summary>
-        /// Listing to a javascript event on the document itself
-        /// </summary>
-        /// <typeparam name="T">The type of the event args for instance MouseEventArgs for mousemove</typeparam>
-        /// <param name="eventName">Name of the DOM event without "on"</param>
-        /// <param name="throotleInterval">The delay between the last time the event occurred and the callback is fired. Set to zero, if no delay is requested</param>
-        /// <param name="callback">The method that is invoked, if the DOM element is fired. Object will be of type T</param>
-        /// <returns>A unique identifier for the event subscription. Should be used to cancel the subscription</returns>
-        Task<Guid> SubscribeGlobal<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicProperties)] T>(string eventName, int throotleInterval, Func<object, Task> callback);
-
-        /// <summary>
-        /// Cancel (unsubscribe) the listening to a DOM event, previous connected by Subscribe
-        /// </summary>
-        /// <param name="key">The unique event identifier</param>
-        /// <returns>true for if the event listener was detached, false if not</returns>
-        Task<bool> Unsubscribe(Guid key);
-    }
-
     public class EventListener : IEventListener, IDisposable
     {
         private readonly IJSRuntime _jsRuntime;
         private readonly DotNetObjectReference<EventListener> _dotNetRef;
-        private bool _disposed = false;
+        private bool _disposed;
 
-        private Dictionary<Guid, (Type eventType, Func<object, Task> callback)> _callbackResolver = new();
+        private readonly Dictionary<Guid, (Type eventType, Func<object, Task> callback)> _callbackResolver = new();
 
         [DynamicDependency(nameof(OnEventOccur))]
         public EventListener(IJSRuntime runtime)
@@ -83,16 +28,7 @@ namespace MudBlazor
         {
             if (_callbackResolver.TryGetValue(key, out var element) && element.callback != null)
             {
-                // Do not move this as static field.
-                // Otherwise, you wil JsonSerializerOptions instances cannot be modified once encapsulated by a JsonSerializerContext exception when using STJ Source Generator.
-                // In net9 remove this and add [JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase, PropertyNameCaseInsensitive = true)] instead.
-                // Alternative we could set TypeInfoResolver for JsonSerializerOptions, but the trimmer will complain that it's unsafe which is not true.
-                var jsonSerializerOptions = new JsonSerializerOptions
-                {
-                    PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-                    PropertyNameCaseInsensitive = true,
-                };
-                var @event = JsonSerializer.Deserialize(eventData, element.eventType, new WebEventJsonContext(jsonSerializerOptions));
+                var @event = JsonSerializer.Deserialize(eventData, element.eventType, WebEventJsonContext.Default);
                 await element.callback.Invoke(@event);
             }
         }
@@ -148,8 +84,6 @@ namespace MudBlazor
             return key;
         }
 
-        #region disposing
-
         public async ValueTask DisposeAsync()
         {
             if (_disposed) { return; }
@@ -200,7 +134,5 @@ namespace MudBlazor
             Dispose(disposing: true);
             GC.SuppressFinalize(this);
         }
-
-        #endregion
     }
 }
