+++ src/MudBlazor/Services/EventManager/EventListenerFactory.cs
@@ -1,0 +1,21 @@
+﻿// Copyright (c) MudBlazor 2021
+// MudBlazor licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.JSInterop;
+
+namespace MudBlazor;
+
+public class EventListenerFactory : IEventListenerFactory
+{
+    private readonly IServiceProvider _provider;
+
+    public EventListenerFactory(IServiceProvider provider)
+    {
+        _provider = provider;
+    }
+
+    public IEventListener Create() =>
+        new EventListener(_provider.GetRequiredService<IJSRuntime>());
+}
