+++ src/MudBlazor/Services/EventManager/WebEventJsonContext.cs
@@ -7,21 +7,19 @@ using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Web;
 using ErrorEventArgs = Microsoft.AspNetCore.Components.Web.ErrorEventArgs;
 
-namespace MudBlazor
-{
-    [JsonSerializable(typeof(EventArgs))]
-    [JsonSerializable(typeof(ChangeEventArgs))]
-    [JsonSerializable(typeof(ClipboardEventArgs))]
-    [JsonSerializable(typeof(DragEventArgs))]
-    [JsonSerializable(typeof(ErrorEventArgs))]
-    [JsonSerializable(typeof(FocusEventArgs))]
-    [JsonSerializable(typeof(KeyboardEventArgs))]
-    [JsonSerializable(typeof(MouseEventArgs))]
-    [JsonSerializable(typeof(PointerEventArgs))]
-    [JsonSerializable(typeof(ProgressEventArgs))]
-    [JsonSerializable(typeof(TouchEventArgs))]
-    [JsonSerializable(typeof(WheelEventArgs))]
-    internal sealed partial class WebEventJsonContext : JsonSerializerContext
-    {
-    }
-}
+namespace MudBlazor;
+
+[JsonSerializable(typeof(EventArgs))]
+[JsonSerializable(typeof(ChangeEventArgs))]
+[JsonSerializable(typeof(ClipboardEventArgs))]
+[JsonSerializable(typeof(DragEventArgs))]
+[JsonSerializable(typeof(ErrorEventArgs))]
+[JsonSerializable(typeof(FocusEventArgs))]
+[JsonSerializable(typeof(KeyboardEventArgs))]
+[JsonSerializable(typeof(MouseEventArgs))]
+[JsonSerializable(typeof(PointerEventArgs))]
+[JsonSerializable(typeof(ProgressEventArgs))]
+[JsonSerializable(typeof(TouchEventArgs))]
+[JsonSerializable(typeof(WheelEventArgs))]
+[JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase, PropertyNameCaseInsensitive = true)]
+internal sealed partial class WebEventJsonContext : JsonSerializerContext;
